added extrareducers file for pokemons
