+++ .firebase/hosting.ZGlzdA.cache
@@ -1,11 +1,8 @@
-404.html,1663927400752,daa499dd96d8229e73235345702ba32f0793f0c8e5c0d30e40e37a5872be57aa
-index.html,1663926546112,26b6fd0679841263f038b829a6ea87460d42882da2ef6383f48ffc6b57406c91
-assets/arrow.0cea72c7.svg,1663926546112,ad9af10e2c3c1b1096c412cb1343c54f4eec557e3d94e1a810646c8914b4bfc2
-assets/facebook.e31cb5c6.svg,1663926546112,2ad85dafc7bfa6ad29f679724cdd3614a84297ebc2df8d5df28efa6a7f37e8af
-assets/index.8f073734.css,1663926546112,961bec6ff91decc44311cba9e8cd7640ba2959e36c7429835e83f4dc4589864b
-assets/instagram.1b433c36.svg,1663926546114,56b6e0c9a8db79cb4558661d44509728ed16b68bdc3efa613336e20b42cf9a29
-assets/pokeball.55134626.png,1663926546112,a3841acf2d5e797f8e409bf494cc9042e427b0d93f04580023f2a3a3597a3696
-assets/telegram.7aeb3bee.svg,1663926546113,2d181be84aa69811503010c082671aff66623cb39229449bdffc835144f89c45
-assets/youtube.1272d1e4.svg,1663926546112,f24b8325e26bf49810454842dd34bda59f8e2be09c07a41b48bdfbbfaf7d6899
-assets/index.44b3e3ea.js,1663926546113,ee1e6965fd7e54d53e08c3a522998bcd594b0094c29be091e7eb6a377f6bced7
-pokeball.png,1663923174200,a3841acf2d5e797f8e409bf494cc9042e427b0d93f04580023f2a3a3597a3696
+index.html,1664095596246,495ffc9e257c82717dac8396edfd41d29232599fc3423bcee74ed269b020b627
+assets/arrow.0cea72c7.svg,1664095596246,ad9af10e2c3c1b1096c412cb1343c54f4eec557e3d94e1a810646c8914b4bfc2
+assets/facebook.e31cb5c6.svg,1664095596246,2ad85dafc7bfa6ad29f679724cdd3614a84297ebc2df8d5df28efa6a7f37e8af
+assets/index.8f073734.css,1664095596247,961bec6ff91decc44311cba9e8cd7640ba2959e36c7429835e83f4dc4589864b
+assets/instagram.1b433c36.svg,1664095596246,56b6e0c9a8db79cb4558661d44509728ed16b68bdc3efa613336e20b42cf9a29
+assets/telegram.7aeb3bee.svg,1664095596246,2d181be84aa69811503010c082671aff66623cb39229449bdffc835144f89c45
+assets/youtube.1272d1e4.svg,1664095596247,f24b8325e26bf49810454842dd34bda59f8e2be09c07a41b48bdfbbfaf7d6899
+assets/index.4c50b232.js,1664095596246,3cf88e891fdbb2bfc0de5bd2c1910f8b47c46d7d4acc91f8482911226115066b
